Refactored Code
I moved all the map logic into a MapControl class. It is also now possible to move into the last room, and to move backwards into the first room.
