+++ Text Adventure/Text Adventure.vcxproj
@@ -144,6 +144,7 @@
   <ItemGroup>
     <ClCompile Include="Character.cpp" />
     <ClCompile Include="gamecore.cpp" />
+    <ClCompile Include="MapControl.cpp" />
   </ItemGroup>
   <ItemGroup>
     <Text Include="map2.txt" />
@@ -154,6 +155,7 @@
     <ClInclude Include="Barren.h" />
     <ClInclude Include="Character.h" />
     <ClInclude Include="Mage.h" />
+    <ClInclude Include="MapControl.h" />
     <ClInclude Include="Rogue.h" />
     <ClInclude Include="Warrior.h" />
   </ItemGroup>
