+++ Text Adventure/maps.txt
@@ -43,8 +43,11 @@ hall
 ***********************_********
 `
 stairwell
-14
+15
 23
+start
+1
+0
 ~
 stairwell
 1
